requirements.txt
websocket-client==0.46.0